% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.rhub.R
\name{groups_h5}
\alias{groups_h5}
\title{HDF5}
\usage{
groups_h5()
}
\value{
string, either "GDALReadOnlyDataset" or "try-error"
}
\description{
Open GDAL HDF5 autotest file.
}
\details{
See mdsumner/hdf5.rhub/data-raw/  for details on getting this file.
}
\examples{
groups_h5()
}

